rename model.utils to model.converter to avoid confused module import
